requirements.txt
Flask
weaviate-client
openai
pdfplumber
markdown